+++ lib/components/z-user-info/z-user-info-view.dart
@@ -1,9 +1,9 @@
 import 'dart:convert';
 import 'dart:typed_data';
+import 'package:airplane_mode_detection/airplane_mode_detection.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:intl/intl.dart';
 import 'package:z_components/api/arquivo/arquivo-service.dart';
 import 'package:z_components/api/arquivo/i-arquivo-service.dart';
 import 'package:z_components/api/endereco/endereco-service.dart';
@@ -75,7 +75,8 @@ class ZUserInfoView extends IView<ZUserInfo> {
     textEditingControllerRua.text = state.widget.userInfo?.logradouro;
     textEditingControllerNumero.text = state.widget.userInfo?.numero;
     if (state.widget.userInfo.dataNascimento != null) {
-      textEditingControllerDataNascimento.text = _montarData(state.widget.userInfo.dataNascimento);
+      textEditingControllerDataNascimento.text =
+          _montarData(state.widget.userInfo.dataNascimento);
     }
   }
 
@@ -84,7 +85,7 @@ class ZUserInfoView extends IView<ZUserInfo> {
     if (state.widget.userInfo.fotoBase64.length > 0) {
       if (state.mounted) {
         state.setState(() {
-          imagemPerfil = base64Decode(state.widget.userInfo.fotoBase64);
+          imagemPerfil =state.widget.userInfo.fotoBase64;
         });
       }
     }
@@ -94,7 +95,11 @@ class ZUserInfoView extends IView<ZUserInfo> {
     if (cep.length == 9) {
       _dialogUtils.showZProgressDialog("Buscando endereço...", 0.5, _globalKey);
 
-      var endereco = await _enderecoService.buscarEnderecoPorCEP(cep);
+      var endereco;
+      String modoAviao = await AirplaneModeDetection.detectAirplaneMode();
+      if (modoAviao == "OFF") {
+        endereco = await _enderecoService.buscarEnderecoPorCEP(cep);
+      }
 
       if (endereco != null) {
         _globalKey.currentState
@@ -111,9 +116,11 @@ class ZUserInfoView extends IView<ZUserInfo> {
           });
         }
       } else {
-        _globalKey.currentState.refresh(
-            1.0, "Não foi possível encontrar o endereço, tenta novamente",
-            success: false);
+        Future.delayed(Duration(milliseconds: 1000), () {
+          _globalKey.currentState.refresh(
+              1.0, "Não foi possível encontrar o endereço, tenta novamente",
+              success: false);
+        });
       }
 
       Future.delayed(new Duration(seconds: 1), () {
@@ -297,46 +304,52 @@ class ZUserInfoView extends IView<ZUserInfo> {
       anexoRGStatus: state.widget.userInfo.anexoRGStatus,
       anexoTituloEleitorStatus: state.widget.userInfo.anexoTituloEleitorStatus,
       idAnexoCartaoAlimentacao: state.widget.userInfo.idAnexoCartaoAlimentacao,
-      idAnexoCartaoAlimentacaoVerso:
-          state.widget.userInfo.idAnexoCartaoAlimentacaoVerso,
+      idAnexoCartaoAlimentacao_Verso:
+          state.widget.userInfo.idAnexoCartaoAlimentacao_Verso,
       idAnexoCartaoContaBancaria:
           state.widget.userInfo.idAnexoCartaoContaBancaria,
-      idAnexoCartaoContaBancariaVerso:
-          state.widget.userInfo.idAnexoCartaoContaBancariaVerso,
+      idAnexoCartaoContaBancaria_Verso:
+          state.widget.userInfo.idAnexoCartaoContaBancaria_Verso,
       idAnexoCartaoValeTransporte:
           state.widget.userInfo.idAnexoCartaoValeTransporte,
-      idAnexoCartaoValeTransporteVerso:
-          state.widget.userInfo.idAnexoCartaoValeTransporteVerso,
+      idAnexoCartaoValeTransporte_Verso:
+          state.widget.userInfo.idAnexoCartaoValeTransporte_Verso,
       idAnexoCarteiraVacinacao: state.widget.userInfo.idAnexoCarteiraVacinacao,
-      idAnexoCarteiraVacinacaoVerso:
-          state.widget.userInfo.idAnexoCarteiraVacinacaoVerso,
+      idAnexoCarteiraVacinacao_Verso:
+          state.widget.userInfo.idAnexoCarteiraVacinacao_Verso,
       idAnexoCertidaoNascimentoCasamento:
           state.widget.userInfo.idAnexoCertidaoNascimentoCasamento,
-      idAnexoCertidaoNascimentoCasamentoVerso:
-          state.widget.userInfo.idAnexoCertidaoNascimentoCasamentoVerso,
+      idAnexoCertidaoNascimentoCasamento_Verso:
+          state.widget.userInfo.idAnexoCertidaoNascimentoCasamento_Verso,
       idAnexoComprovanteEndereco:
           state.widget.userInfo.idAnexoComprovanteEndereco,
-      idAnexoComprovanteEnderecoVerso:
-          state.widget.userInfo.idAnexoComprovanteEnderecoVerso,
+      idAnexoComprovanteEndereco_Verso:
+          state.widget.userInfo.idAnexoComprovanteEndereco_Verso,
       idAnexoCPF: state.widget.userInfo.idAnexoCPF,
-      idAnexoCPFVerso: state.widget.userInfo.idAnexoCPFVerso,
+      idAnexoCPF_Verso: state.widget.userInfo.idAnexoCPF_Verso,
       idAnexoCTPS: state.widget.userInfo.idAnexoCTPS,
-      idAnexoCTPSVerso: state.widget.userInfo.idAnexoCTPSVerso,
+      idAnexoCTPS_Verso: state.widget.userInfo.idAnexoCTPS_Verso,
       idAnexoEscolaridade: state.widget.userInfo.idAnexoEscolaridade,
-      idAnexoEscolaridadeVerso: state.widget.userInfo.idAnexoEscolaridadeVerso,
+      idAnexoEscolaridade_Verso:
+          state.widget.userInfo.idAnexoEscolaridade_Verso,
       idAnexoPIS: state.widget.userInfo.idAnexoPIS,
-      idAnexoPISVerso: state.widget.userInfo.idAnexoPISVerso,
+      idAnexoPIS_Verso: state.widget.userInfo.idAnexoPIS_Verso,
       idAnexoRG: state.widget.userInfo.idAnexoRG,
-      idAnexoRGVerso: state.widget.userInfo.idAnexoRGVerso,
+      idAnexoRG_Verso: state.widget.userInfo.idAnexoRG_Verso,
       idAnexoTituloEleitor: state.widget.userInfo.idAnexoTituloEleitor,
-      idAnexoTituloEleitorVerso:
-          state.widget.userInfo.idAnexoTituloEleitorVerso,
+      idAnexoTituloEleitor_Verso:
+          state.widget.userInfo.idAnexoTituloEleitor_Verso,
     );
 
     _dialogUtils.showZProgressDialog(
         "Salvando informações...", 0.7, _globalKey);
 
-    var res = await _userInfoService.editarInformacoes(userInfo);
+    var res = false;
+    String modoAviao = await AirplaneModeDetection.detectAirplaneMode();
+
+    if (modoAviao == "OFF") {
+      res = await _userInfoService.editarInformacoes(userInfo);
+    }
 
     if (res) {
       userInfo.atualizado = true;
@@ -344,8 +357,9 @@ class ZUserInfoView extends IView<ZUserInfo> {
       userInfo.atualizado = false;
     }
 
-    _globalKey.currentState.refresh(1.0, "Pronto", success: true);
-
+    Future.delayed(Duration(milliseconds: 1000), () {
+      _globalKey.currentState.refresh(1.0, "Pronto", success: true);
+    });
     Future.delayed(new Duration(seconds: 1), () {
       _dialogUtils.dismiss();